OrcaKeep - a simple accommodation booking platform
a simple web project - a booking platform that helps users/clients search and secure their favourite accommodation/hotels as quickly as possible. This took inspiration from AirBnb. Click on the search button next to the search bar without input will show all available options. The website also has a registration form and a log-in dropdown form. The social media following links on the footer. This one for Couch to Coder 2024 and using Microsoft VS Code as IDE